+++ bench/opt2026/setup_cluster_environment.jl
@@ -14,8 +14,11 @@ isempty(source) &&
 
 Pkg.develop(PackageSpec(path=source))
 Pkg.add([
+    PackageSpec(name="Aqua", version="0.8"),
+    PackageSpec(name="Convex", version="0.16"),
     PackageSpec(name="DoubleFloats", version="1"),
     PackageSpec(name="JLD2"),
+    PackageSpec(name="JSON"),
     PackageSpec(name="JuMP", version="1"),
     PackageSpec(name="MultiFloats", version="3"),
     PackageSpec(name="StableRNGs"),